handle double values for numeric fields

+++ lib/features/bp/editor.dart
@@ -411,7 +411,8 @@ class _BPEditorState extends State<BPEditor> {
           keyboardType: TextInputType.number,
           onChanged: (value) {
             records[index] = record.copyWith(
-              systolic: int.tryParse(value) ?? 0,
+              systolic: double.tryParse(value) ??
+                  0, // Use double instead of int for potential readings with decimal places.
             );
           },
         )),
@@ -420,7 +421,7 @@ class _BPEditorState extends State<BPEditor> {
           keyboardType: TextInputType.number,
           onChanged: (value) {
             records[index] = record.copyWith(
-              diastolic: int.tryParse(value) ?? 0,
+              diastolic: double.tryParse(value) ?? 0,
             );
           },
         )),
@@ -429,7 +430,7 @@ class _BPEditorState extends State<BPEditor> {
           keyboardType: TextInputType.number,
           onChanged: (value) {
             records[index] = record.copyWith(
-              heartRate: int.tryParse(value) ?? 0,
+              heartRate: double.tryParse(value) ?? 0,
             );
           },
         )),
